refactor: move katex component out of lib to reduce bundle
