+++ src/content/blogpost/radix/components/index.ts
@@ -1,3 +1,4 @@
 export { default as FormNumerals } from "./form-numerals/FormNumerals.svelte";
 export { default as Number } from "./number/Number.svelte";
 export { default as Numerals } from "./numerals/Numerals.svelte";
+export { default as MathExp } from "./math.svelte";
