+++ src/content/blogpost/radix/interactive/number-into-decimal/NumberIntoDecimal.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
 	import { getContext } from "svelte";
 	import { radix } from "@ericrovell/radix";
-	import { MathExp } from "@components";
-	import { FormNumerals, Number, Numerals } from "../../components";
+	import { FormNumerals, Number, Numerals, MathExp } from "../../components";
 	import { decoder } from "../../util";
 	import styles from "./number-into-decimal.module.css";
