+++ src/content/blogpost/where-do-aperture-numbers-come-from/components/math.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+	import katex from "katex";
+
+	export let math: string;
+	export let inline = false;
+
+	$: rendered = katex.renderToString(math, {
+		throwOnError: false,
+		displayMode: inline
+	});
+</script>
+
+{@html rendered}
