--- src/lib/components/math/index.ts
@@ -1,1 +1,0 @@
-export { default as MathExp } from "./MathExp.svelte";
